requirements.txt
transformers
datasets
accelerate
bitsandbytes
trl
peft
evaluate
wandb
pyyaml
torch
